Move `pokemon-[form-]level-moves.ts`
