+++ src/data/pokemon-forms.ts
@@ -1,6 +1,6 @@
 // -- start tsdoc imports --
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import type { FORM_CHANGE_MOVE } from "#app/data/balance/pokemon-level-moves";
+import type { FORM_CHANGE_MOVE } from "#app/data/pokemon-level-moves";
 import type { FormChangePhase } from "#app/phases/form-change-phase";
 import type { QuietFormChangePhase } from "#app/phases/quiet-form-change-phase";
 /* eslint-enable @typescript-eslint/no-unused-vars */
