+++ src/data/pokemon-species-form.ts
@@ -1,22 +1,22 @@
 import type { StarterMoveset } from "#app/@types/StarterData";
 import type { AnySound } from "#app/audio-manager";
-import { speciesEggMoves } from "#app/data/egg-moves";
-import { pokemonPreEvolutions } from "#app/data/pokemon-pre-evolutions";
-import { type LevelMoves, pokemonSpeciesLevelMoves } from "#app/data/balance/pokemon-level-moves";
-import { pokemonFormLevelMoves } from "./balance/pokemon-form-level-moves";
+import { type LevelMoves, pokemonSpeciesLevelMoves } from "#app/data/pokemon-level-moves";
 import { speciesStarterCosts } from "#app/data/balance/starters";
+import { speciesEggMoves } from "#app/data/egg-moves";
 import { noRandomSpeciesSpawn } from "#app/data/no-random-species-spawn";
-import type { PokemonForm } from "./pokemon-form";
-import { variantData, type VariantSet, type Variant } from "#app/data/variant";
+import { pokemonFormLevelMoves } from "#app/data/pokemon-form-level-moves";
+import { pokemonPreEvolutions } from "#app/data/pokemon-pre-evolutions";
+import { type Variant, variantData, type VariantSet } from "#app/data/variant";
 import { globalScene } from "#app/global-scene";
 import { getPokemonSpecies } from "#app/utils/pokemon-species-utils";
 import { AbilityId } from "#enums/ability-id";
+import type { ElementalType } from "#enums/elemental-type";
 import { PokemonRegion } from "#enums/pokemon-regions";
-import { SpeciesId } from "#enums/species-id";
 import { SpeciesFormKey } from "#enums/species-form-key";
+import { SpeciesId } from "#enums/species-id";
 import type { Stat } from "#enums/stat";
 import { argbFromRgba, QuantizerCelebi, rgbaFromArgb } from "@material/material-color-utilities";
-import type { ElementalType } from "#enums/elemental-type";
+import type { PokemonForm } from "./pokemon-form";
 
 //#region Types
