+++ src/phases/evolution-phase.ts
@@ -4,7 +4,7 @@ import { type FormChangePhase } from "#app/phases/form-change-phase";
 // -- end tsdoc imports --
 
 import type { AnySound } from "#app/audio-manager";
-import { EVOLVE_MOVE } from "#app/data/balance/pokemon-level-moves";
+import { EVOLVE_MOVE } from "#app/data/pokemon-level-moves";
 import type { SpeciesFormEvolution } from "#app/data/pokemon-evolutions";
 import type { PlayerPokemon, Pokemon } from "#app/field/pokemon";
 import { globalScene } from "#app/global-scene";
